requirements.txt
# Core ML Libraries
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0

# Data Processing
nltk>=3.6.0

# Configuration and Logging
PyYAML>=5.4.0

# DVC
dvc>=2.0.0

# Production-grade dependencies
joblib>=1.0.0

# Development and Testing
pytest>=6.0.0
pytest-cov>=2.12.0
black>=21.0.0
flake8>=3.9.0
mypy>=0.910

# Jupyter (optional)
jupyter>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Cookiecutter template dependencies
click>=8.0.0
python-dotenv>=0.19.01
